requirements.txt
fastapi
uvicorn[standard]
selenium
python-dotenv
slowapi
